Refactor slide embedding and add batch rendering
